+++ src/getStepStartStates.test.ts
@@ -195,6 +195,9 @@ describe('getStepStartStates', () => {
       [SecuritySteps.PRICING_CONFIGURATIONS]: { disabled: true },
       [MonitorSteps.MONITOR_LOG_PROFILES]: { disabled: true },
       [MonitorSteps.MONITOR_ACTIVITY_LOG_ALERTS]: { disabled: true },
+      [MonitorSteps.MONITOR_ACTIVITY_LOG_ALERT_SCOPE_RELATIONSHIPS]: {
+        disabled: true,
+      },
     });
   });
 
@@ -288,6 +291,9 @@ describe('getStepStartStates', () => {
       [SecuritySteps.PRICING_CONFIGURATIONS]: { disabled: true },
       [MonitorSteps.MONITOR_LOG_PROFILES]: { disabled: true },
       [MonitorSteps.MONITOR_ACTIVITY_LOG_ALERTS]: { disabled: true },
+      [MonitorSteps.MONITOR_ACTIVITY_LOG_ALERT_SCOPE_RELATIONSHIPS]: {
+        disabled: true,
+      },
     });
   });
 
@@ -381,6 +387,9 @@ describe('getStepStartStates', () => {
       [SecuritySteps.PRICING_CONFIGURATIONS]: { disabled: false },
       [MonitorSteps.MONITOR_LOG_PROFILES]: { disabled: false },
       [MonitorSteps.MONITOR_ACTIVITY_LOG_ALERTS]: { disabled: false },
+      [MonitorSteps.MONITOR_ACTIVITY_LOG_ALERT_SCOPE_RELATIONSHIPS]: {
+        disabled: false,
+      },
     });
   });
 });