+++ test/helpers/getMockEntity.ts
@@ -3,6 +3,7 @@ import { v4 as uuid } from 'uuid';
 import { IntegrationConfig } from '../../src/types';
 import { createResourceGroupEntity } from '../../src/steps/resource-manager/resources/converters';
 import { createAzureWebLinker } from '../../src/azure';
+import { createSubscriptionEntity } from '../../src/steps/resource-manager/subscriptions/converters';
 
 const webLinker = createAzureWebLinker(undefined);
 
@@ -21,5 +22,11 @@ export function getMockResourceGroupEntity(resourceGroupName: string) {
     name: resourceGroupName,
     location: 'useast1',
     id: 'RESOURCE_GROUP_ID',
+  });
+}
+
+export function getMockSubscriptionEntity(config: IntegrationConfig) {
+  return createSubscriptionEntity(webLinker, {
+    id: `/subscriptions/${config.subscriptionId}`,
   });
 }